fix(???): Implement `draw.line` function, updating everything as f64 and lerping w/ @as(f32)
I was having trouble with drawing straight lines in the terminal, (0..width)
obviously it has something to do with numerical precision, but for some
reason this fixed it:
```
src/draw.zig:200:42| const interpolation = a.lerp(@as(f32, @floatFromInt(t)) / 400, b);
```
having the lerp be @as(f32) and everything else as f64 made it work,
what the eff???

TODO: Test why this is happening further, most likely
float precision / rounding issues here

+++ src/braille.zig
@@ -17,14 +17,14 @@ pub const Plotter = struct {
         self.buffer.clearRetainingCapacity();
     }
 
-    pub fn get(self: *Plotter, x: f32, y: f32) ![3]u8 {
+    pub fn get(self: *Plotter, x: f64, y: f64) ![3]u8 {
         const key = Key{ @intFromFloat(x), @intFromFloat(y) };
         if (try self.buffer.get(key)) |entry| {
             return BraillePoint(entry.value_ptr.*);
         } else return null;
     }
 
-    pub fn plot(self: *Plotter, x: f32, y: f32) !void {
+    pub fn plot(self: *Plotter, x: f64, y: f64) !void {
         const key = Key{ @intFromFloat(x), @intFromFloat(y) };
         const sx = @trunc(@mod(x, 1) * 2);
         const sy = @trunc(@mod(y, 1) * 4);
